requirements.txt
Flask
flask-cors
numpy
tensorflow-cpu
